% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ca1115}
\alias{ca1115}
\title{Medicaid Income Eligibility Limits for Childless Adults, 2011-2015}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 255 rows and 6 columns.}
\source{
/url{http://kff.org/medicaid/state-indicator/medicaid-income-eligibility-limits-for-other-non-disabled-adults-2011-2015/}
}
\usage{
ca1115
}
\description{
Medicaid Income Eligibility Limits for Childless Adults, 2011-2015
}
\keyword{datasets}

