% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{childless_adults}
\alias{childless_adults}
\title{Medicaid Income Eligibility Limits for Childless Adults, 2011-2018}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 408 rows and 6 columns.}
\source{
\href{https://www.kff.org/medicaid/state-indicator/medicaid-income-eligibility-limits-for-other-non-disabled-adults/}{Medicaid Income Eligibility Limits for Other Non-Disabled Adults, 2011-2018}
}
\usage{
childless_adults
}
\description{
Medicaid Income Eligibility Limits for Childless Adults, 2011-2018
}
\keyword{datasets}
