% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{children}
\alias{children}
\title{Medicaid/CHIP Income Eligibility Limits for Children, 2000-2015}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3060 rows and 8 columns.}
\source{
/url{http://kff.org/data-collection/trends-in-medicaid-income-eligibility-limits/}
}
\usage{
children
}
\description{
Medicaid/CHIP Income Eligibility Limits for Children, 2000-2015
}
\keyword{datasets}

