% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melig-package.R
\docType{package}
\name{melig-package}
\alias{melig}
\alias{melig-package}
\title{melig: Trends in Medicaid Income Eligibility Limits}
\description{
This R data package contains data on the trends in Medicaid income eligibility limits for children, childless adults, parents, and pregnant women. The dataset is sourced from the Kaiser Family Foundation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jjchern/melig}
  \item Report bugs at \url{https://github.com/jjchern/melig/issues}
}

}
\author{
\strong{Maintainer}: JJ Chen \email{jiajia.chern@gmail.com} (\href{https://orcid.org/0000-0001-8482-8398}{ORCID})

}
\keyword{internal}
