% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pw0315}
\alias{pw0315}
\title{Medicaid/CHIP Income Eligibility Limits for Pregnant Women, 2003-2015}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 612 rows and 6 columns.}
\source{
/url{http://kff.org/medicaid/state-indicator/medicaid-and-chip-income-eligibility-limits-for-pregnant-women-2003-2015/}
}
\usage{
pw0315
}
\description{
Medicaid/CHIP Income Eligibility Limits for Pregnant Women, 2003-2015
}
\keyword{datasets}

