% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{childless_adults}
\alias{childless_adults}
\title{Medicaid Income Eligibility Limits for Childless Adults, 2011-2016}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 306 rows and 6 columns.}
\source{
/url{http://kff.org/medicaid/state-indicator/medicaid-income-eligibility-limits-for-other-non-disabled-adults-2011-2015/}
}
\usage{
childless_adults
}
\description{
Medicaid Income Eligibility Limits for Childless Adults, 2011-2016
}
\keyword{datasets}

