% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melig-package.R
\docType{package}
\name{melig-package}
\alias{melig}
\alias{melig-package}
\title{melig: Trends in Medicaid Income Eligibility Limits}
\description{
An R data package contains data for trends in Medicaid income eligibility limits. 
    Source: The The Kaiser Family Foundation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jjchern/melig}
  \item Report bugs at \url{https://github.com/jjchern/melig/issues}
}

}
\author{
\strong{Maintainer}: JJ Chen \email{jiajia.chern@gmail.com} (0000-0001-8482-8398)

}
\keyword{internal}
