% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pa0215}
\alias{pa0215}
\title{Medicaid Income Eligibility Limits for Parents, 2002-2015}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 663 rows and 6 columns.}
\source{
/url{http://kff.org/medicaid/state-indicator/medicaid-income-eligibility-limits-for-parents-2002-2015}

/url{http://kff.org/medicaid/state-indicator/medicaid-income-eligibility-limits-for-other-non-disabled-adults-2011-2015/}

/url{http://kff.org/medicaid/state-indicator/medicaid-and-chip-income-eligibility-limits-for-pregnant-women-2003-2015/}
}
\usage{
pa0215

pa0215

pa0215
}
\description{
Medicaid Income Eligibility Limits for Parents, 2002-2015

Medicaid Income Eligibility Limits for Childless Adults, 2011-2015

Medicaid/CHIP Income Eligibility Limits for Pregnant Women, 2003-2015
}
\keyword{datasets}

